+++ .gitmodules
@@ -1,3 +1,3 @@
 [submodule "packages/ckb-wallet/udtScripts"]
-	path = packages/ckb-wallet/udtScripts
+	path = packages/ckb-sdk-wallet/udtScripts
 	url = https://github.com/nervosnetwork/ckb-ruby-scripts/
