--- packages/ckb-core/__tests__/ckb-core.test.js
@@ -1,9 +1,0 @@
-const ckbCore = require('..')
-
-describe('ckb-core', () => {
-  'use strict'
-
-  it.skip('needs tests', () => {
-    // TODO
-  })
-})
