--- packages/ckb-core/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@ckb-sdk/core",
-  "version": "0.0.0",
-  "description": "> TODO: description",
-  "author": "Nervos <dev@nervos.org>",
-  "homepage": "https://github.com/nervosnetwork/ckb-sdk-js#readme",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "publishConfig": {
-    "registry": "http://registry.npmjs.org/",
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nervosnetwork/ckb-sdk-js.git"
-  },
-  "scripts": {
-    "tsc": "tsc"
-  },
-  "bugs": {
-    "url": "https://github.com/nervosnetwork/ckb-sdk-js/issues"
-  },
-  "dependencies": {
-    "@ckb-sdk/rpc": "^0.0.0",
-    "@ckb-sdk/types": "^0.0.0",
-    "@ckb-sdk/utils": "^0.0.0",
-    "@ckb-sdk/wallet": "^0.0.0"
-  },
-  "devDependencies": {
-    "@types/crypto-js": "^3.1.43"
-  }
-}
