--- packages/ckb-core/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "types": ["@ckb-sdk/types", "@ckb-sdk/utils"]
-  },
-  "include": ["./src"]
-}
