--- packages/ckb-rpc/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@ckb-sdk/rpc",
-  "version": "0.0.0",
-  "description": "@ckb-sdk rpc module",
-  "author": "Nervos <dev@nervos.org>",
-  "homepage": "https://github.com/nervosnetwork/ckb-sdk-js/packages/ckb-rpc#readme",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "publishConfig": {
-    "registry": "http://registry.npmjs.org/",
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nervosnetwork/ckb-sdk-js.git"
-  },
-  "scripts": {
-    "tsc": "tsc",
-    "test": "../../node_modules/.bin/jest",
-    "doc": "../../node_modules/.bin/typedoc --out docs --mode modules --includeDeclarations --excludeExternals --ignoreCompilerErrors --theme default",
-    "test:watch": "../../node_modules/.bin/jest --watch"
-  },
-  "bugs": {
-    "url": "https://github.com/nervosnetwork/ckb-sdk-js/issues"
-  },
-  "dependencies": {
-    "@ckb-sdk/utils": "^0.0.0",
-    "axios": "^0.18.0"
-  },
-  "devDependencies": {
-    "@ckb-sdk/types": "^0.0.0",
-    "@types/crypto-js": "^3.1.43"
-  }
-}
