--- packages/ckb-rpc/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "target": "es6",
-    "lib": ["es6", "dom"],
-    "outDir": "./lib",
-    "types": ["@ckb-sdk/types", "@ckb-sdk/utils"]
-  },
-  "include": ["./src"]
-}
