+++ packages/ckb-sdk-core/__tests__/ckb-core.test.js
@@ -1,0 +1,9 @@
+// const ckbCore = require('..')
+
+describe('ckb-core', () => {
+  // 'use strict'
+
+  it.skip('needs tests', () => {
+    // TODO
+  })
+})
