+++ packages/ckb-sdk-rpc/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "target": "es6",
+    "lib": ["es6", "dom"],
+    "outDir": "./lib",
+    "types": ["@nervosnetwork/ckb-types", "@nervosnetwork/ckb-sdk-utils"]
+  },
+  "include": ["./src"]
+}
