+++ packages/ckb-sdk-utils/__tests__/ckb-utils.test.js
@@ -1,0 +1,7 @@
+// const ckbUtils = require('..')
+
+describe('ckb-utils', () => {
+  it.skip('needs tests', () => {
+    // TODO
+  })
+})
