+++ packages/ckb-sdk-wallet/__tests__/ckb-wallet.test.js
@@ -1,0 +1,7 @@
+// const ckbWallet = require('..')
+
+describe('ckb-wallet', () => {
+  it.skip('needs tests', () => {
+    // TODO
+  })
+})
