+++ packages/ckb-sdk-wallet/package.json
@@ -1,14 +1,18 @@
 {
-  "name": "@nervosnetwork/ckb-types",
+  "name": "@nervosnetwork/ckb-sdk-wallet",
   "version": "0.0.0",
   "description": "> TODO: description",
   "author": "Nervos <dev@nervos.org>",
   "homepage": "https://github.com/nervosnetwork/ckb-sdk-js#readme",
   "license": "MIT",
-  "types": "index.d.ts",
+  "main": "lib/index.js",
   "directories": {
+    "lib": "lib",
     "test": "__tests__"
   },
+  "files": [
+    "lib"
+  ],
   "publishConfig": {
     "registry": "http://registry.npmjs.org/"
   },
@@ -16,13 +20,20 @@
     "type": "git",
     "url": "git+https://github.com/nervosnetwork/ckb-sdk-js.git"
   },
+  "scripts": {
+    "tsc": "tsc",
+    "test": "../../node_modules/.bin/jest",
+    "test:watch": "../../node_modules/.bin/jest --watch"
+  },
   "bugs": {
     "url": "https://github.com/nervosnetwork/ckb-sdk-js/issues"
   },
-  "scripts": {
-    "doc": "../../node_modules/.bin/typedoc --out docs ./index.d.ts --mode modules --includeDeclarations --excludeExternals --ignoreCompilerErrors --theme default --readme README.md"
+  "dependencies": {
+    "@nervosnetwork/ckb-sdk-rpc": "^0.0.0",
+    "@nervosnetwork/ckb-sdk-utils": "^0.0.0"
   },
   "devDependencies": {
+    "@nervosnetwork/ckb-types": "^0.0.0",
     "@types/crypto-js": "^3.1.43"
   }
 }
