+++ packages/ckb-sdk-wallet/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./lib",
+    "types": ["node", "@nervosnetwork/ckb-sdk-rpc", "@nervosnetwork/ckb-sdk-utils"]
+  },
+  "include": ["./src"]
+}
