--- packages/ckb-utils/__tests__/ckb-utils.test.js
@@ -1,9 +1,0 @@
-const ckbUtils = require('..')
-
-describe('ckb-utils', () => {
-  'use strict'
-
-  it.skip('needs tests', () => {
-    // TODO
-  })
-})
