--- packages/ckb-utils/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@ckb-sdk/utils",
-  "version": "0.0.0",
-  "description": "> TODO: description",
-  "author": "Nervos <dev@nervos.org>",
-  "homepage": "https://github.com/nervosnetwork/ckb-sdk-js#readme",
-  "license": "MIT",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib"
-  ],
-  "publishConfig": {
-    "registry": "http://registry.npmjs.org/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nervosnetwork/ckb-sdk-js.git"
-  },
-  "scripts": {
-    "tsc": "tsc"
-  },
-  "bugs": {
-    "url": "https://github.com/nervosnetwork/ckb-sdk-js/issues"
-  },
-  "dependencies": {
-    "@ckb-sdk/types": "^0.0.0",
-    "bech32": "^1.1.3",
-    "bs58check": "^2.1.2",
-    "eth-lib": "^0.2.8",
-    "randombytes": "^2.0.6",
-    "safe-buffer": "^5.1.2",
-    "sha3": "^2.0.0",
-    "tiny-secp256k1": "^1.0.1",
-    "typeforce": "^1.18.0",
-    "utf8": "^3.0.0",
-    "wif": "^2.0.6"
-  },
-  "devDependencies": {
-    "@types/bitcoinjs-lib": "^4.0.0",
-    "@types/tiny-secp256k1": "^1.0.0",
-    "@types/utf8": "^2.1.6"
-  }
-}
