--- packages/ckb-utils/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "target": "es6",
-    "outDir": "./lib",
-    "typeRoots": ["./types", "./node_modules/@types"],
-    "types": ["@ckb-sdk/types", "node", "@types/tiny-secp256k1", "sha3"]
-  },
-  "include": ["./src"],
-  "exclude": ["node_modules", "typings", "./types"]
-}
