--- packages/ckb-wallet/__tests__/ckb-wallet.test.js
@@ -1,9 +1,0 @@
-const ckbWallet = require('..')
-
-describe('ckb-wallet', () => {
-  'use strict'
-
-  it.skip('needs tests', () => {
-    // TODO
-  })
-})
