--- packages/ckb-wallet/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./lib",
-    "types": ["node", "@ckb-sdk/rpc", "@ckb-sdk/utils"]
-  },
-  "include": ["./src"]
-}
